Implemented ambient occlusion (SSAO and HBAO)
Added ambient occlusion rendering pass
Added SSAO pass and SSAO blur shaders
Added HBAO pass shaders
Added HBAO blur shaders that are split into horizontal and vertical passes, and takes into account fragments depth (so to not blur edges)
Added AmbientOcclusionData container that holds all the AO settings, in Containers
Added AODataSet container that holds the AO shader settings and gets sent to the GPU as Uniform block, in GraphicsDataSets, UnfirormData
Added ambient occlution data to RendererScene, so each map can load and have its own AO settings, that get sent to the renderer
Added a transpose inverse of view matrix calculation on the CPU and sending it as a uniform to shaders, in RendererFrontend, ShaderUniforms, ShaderUniformUpdater
Added settings for ambient occlusion, in EditorWindow
Added more texture data types and texture data formats, in EditorWindow, CommonDefinitions
Added more uniforms in ShaderUniforms, ShaderUniformUpdater
Added a way to set magnification and minification filtering for each texture separately, in CommandBuffer, RendererBackend, RendererFrontend
Added texture filtering type enum in CommonDefinitions
Added vertex tangent and bitangent calculation flags for Assimp, instead of calculating them manually while loading models, in ModelLoader
Added more Config variables
Added more PropertyIDs
Added more ErrorCodes
Incremented engine version to 0.1.3 and commit date to 2023-12-24
Fixed a bug of error window popup showing text color tags in plain text, in ErrorHandler

+++ Dependencies/include/ImGuiColorTextEdit/TextEditor.h
@@ -43,7 +43,7 @@ public:
 	inline bool IsShowLineNumbersEnabled() const { return mShowLineNumbers; }
 	inline void SetShortTabsEnabled(bool aValue) { mShortTabs = aValue; }
 	inline bool IsShortTabsEnabled() const { return mShortTabs; }
-	inline int GetLineCount() const { return mLines.size(); }
+	inline int GetLineCount() const { return (int)mLines.size(); }
 	inline bool IsOverwriteEnabled() const { return mOverwrite; }
 	void SetPalette(PaletteId aValue);
 	PaletteId GetPalette() const { return mPaletteId; }
